+++ JUN_All/tools/A00290_BSTool/app/core/__init__.py
@@ -1,5 +1,6 @@
 from .edit_bs_manager import EditBSManager
 from .base_shape_manager import BaseShapeManager
+from .shape_editor_manager import ShapeEditorManager
 from . import blendshape_utils
 
-__all__ = ["EditBSManager", "BaseShapeManager", "blendshape_utils"]
+__all__ = ["EditBSManager", "BaseShapeManager", "ShapeEditorManager", "blendshape_utils"]